Fix external repository location
The XLDeploy server config allows you to specify an alternative
repository location. But when you specify an absolute path it will not
work. Reason is that the entry in the deployit.conf then needs to be
prefixed with file://. Putting that as a parameter however will not
work as it will break the xldeploy init script's creates check.
We needed to put the repository outside of the server install folder to
allow puppet to modify the config of the server (including upgrading to
a new version) without reinitializing the repo.
To fix this the creates check on the init now uses the
jcr_repository_path variable. And in the ini_setting a regexp
substitution has been added to check if the path starts with a / and if
so replaces it with file:///