Removed puppet file server dependency on puppetfiles_xldeploy_source
The puppetfiles_xldeploy_source variable assumed the specified location
to be on the puppet file server. This prevented getting files from a
different location.
In our case we download the required zips from a repository server and
want to use those files. Be removing the puppet:/// from the install.pp
we can now specify the protocol in the parameter.